refactor(proxy): move Anthropic model-list formatter into llms/anthropic/common_utils
