+++ litellm/proxy/proxy_server.py
@@ -8275,8 +8275,10 @@ async def model_list(
     from litellm.proxy.management_endpoints.common_utils import (
         _user_has_admin_privileges,
     )
-    from litellm.proxy.utils import (
+    from litellm.llms.anthropic.common_utils import (
         create_anthropic_model_list_response,
+    )
+    from litellm.proxy.utils import (
         create_model_info_response,
         get_available_models_for_user,
     )
